requirements.txt
# web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# kafka
aiokafka==0.11.0

# database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# cache
redis[hiredis]==5.2.1

# ML - phase 1
scikit-learn==1.6.1
numpy==2.2.1
pandas==2.2.3
joblib==1.4.2

# ML - phase 2
torch==2.6.0

# config & utilities
pydantic-settings==2.7.1
python-json-logger==3.2.1
httpx==0.28.1

# testing
pytest==8.3.4
pytest-asyncio==0.25.0
